+++ ledger/catchupaccessor_test.go
@@ -212,7 +212,7 @@ func verifyStateProofVerificationCatchupAccessor(t *testing.T, targetData []ledg
 
 	var trackedStateProofVerificationContext []ledgercore.StateProofVerificationContext
 	err = l.trackerDBs.Rdb.Atomic(func(ctx context.Context, tx *sql.Tx) error {
-		dbData, err := store.StateProofVerification(ctx, tx)
+		dbData, err := store.GetAllStateProofVerification(ctx, tx)
 		trackedStateProofVerificationContext = dbData
 		return err
 	})
